Implement tracking of deleted items in snapshot API
Now snapshot requests can include information on resource tree nodes
deleted since the last flush of the tree.  As part of this, formatters
can request multiple sequential scans of the resource tree filtered by
node state as part of a snapshot operation.  As an example, this allows
the existing nodes to be considered in one pass, and the deleted nodes
to be considered in another.

Each node now has three states indicating its lifetime:
 * "New," indicating it has never been captured by a previous snapshot.
 * "Deleted," indicating the node has been deleted.  Deleted nodes are
   only provided if deletion tracking is turned on.
 * "Normal," indicating the node is in the regular portion of its
   existence.

A pair of periodically added and removed resources have been added to
the actuator demo program to assist with testing.

Resolves: LE-14968
Change-Id: I979ad3d83ae3000d4c820a6948ae2a651e4bbb18