Add a wrapper script for running DH in GDB
This sets up the environment and runs the Data Hub in GDB, within a
tmux session which can be attached to.  Some initial data is also
populated in the Data Hub.  The startup takes a few seconds to
complete.  Use "make stop" to terminate the environment.  The GDB
session will still need to be closed manually.

Resolves: LE-14968
Change-Id: I5e15c2faee520cbce1b3c8b5f59e4182207bdfee